protected video route

+++ controller/videoController.js
@@ -19,7 +19,6 @@ export const search = async (req, res) => {
     let videos = [];
     try {
         videos = await Video.find({ title: { $regex: searchingBy, $options: 'i' } });
-        console.log(videos);
     } catch (error) {
         console.log(error);
     }
@@ -51,7 +50,6 @@ export const videoDetail = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id).populate('creator');
-        console.log(video);
         res.render('videoDetail', { pageTitle: `${video.title}`, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -65,7 +63,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -89,7 +91,12 @@ export const deleteVideo = async (req, res) => {
         params: { id },
     } = req;
     try {
-        await Video.findOneAndRemove({ _id: id });
+        const video = await Video.findById(id);
+        if (video.creator !== id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch (error) {
         console.log(error);
     }
